many changes
new create script, new insert data script, new select from where query, added to Q9

+++ SocialMediaDatabase.sql
@@ -143,7 +143,7 @@ CREATE TABLE participates (
 
 CREATE TABLE advertisement (
 	pid integer NOT NULL,
-	price integer,
+	price decimal(7,2),
     PRIMARY KEY(pid),
     FOREIGN KEY (pid) REFERENCES post(pid)    
 );
